flask-otlp/requirements.txt
requests==2.27.1
blinker==1.6.2
flask==3.0.0
flask-cors==3.0.10
gunicorn==20.1.0
pg8000==1.16.6
psycopg2-binary==2.9.10
python-dotenv==0.12.0
pytz==2020.4
sentry-sdk[statsig,opentelemetry-otlp]>=2.48.0
sqlalchemy==1.4.49
redis==5.0.1
Werkzeug==3.0.3
flask_caching==2.3.0
openai==1.52.2
tiktoken==0.8.0
celery==5.4.0
statsig==0.57.1
opentelemetry-api>=1.38.0
opentelemetry-sdk>=1.38.0
opentelemetry-semantic-conventions>=0.59b0
opentelemetry-instrumentation-flask>=0.59b0
opentelemetry-instrumentation-sqlalchemy>=0.59b0
opentelemetry-instrumentation-redis>=0.59b0
opentelemetry-instrumentation-requests>=0.59b0
